fix(web): resolve CI failings and refactor MainContent
- Refactored MainContent monolithic component into sub-components (MainHeader, SearchSection, MetadataBar) to reduce cyclomatic complexity and JSX nesting.
- Converted functional components to arrow functions to address DeepSource JS-0067.
- Removed unused `mobileMenuOpen` variable from MainContent.
- Improved accessibility and semantic HTML by using <header>, <section>, and <main> tags.
- Fixed 'Clear' button visibility logic to ensure it remains available when results are present.
- Verified with full test suite and quality gate.
