+++ web/app/components/MainContent.tsx
@@ -1,8 +1,10 @@
 "use client";
 
-import Link from "next/link";
+import { MainHeader } from "@/app/components/MainHeader";
+import { SearchSection } from "@/app/components/SearchSection";
+import { MetadataBar } from "@/app/components/MetadataBar";
 import ResultCard from "@/app/components/ResultCard";
-import { ProviderResult } from "@/lib/results";
+import type { ProviderResult } from "@/lib/results";
 
 interface MainContentProps {
   mobileMenuOpen: boolean;
@@ -35,186 +37,6 @@ interface MainContentProps {
   copied: boolean;
   isUrl: boolean;
 }
-
-const MainHeader = ({ setMobileMenuOpen }: { setMobileMenuOpen: (open: boolean) => void }) => (
-  <div className="border-b-2 border-border-muted p-2 flex items-center justify-between min-h-[44px]">
-    <div className="flex items-center gap-2">
-      <button
-        onClick={() => setMobileMenuOpen(true)}
-        className="lg:hidden p-2 text-text-muted hover:text-foreground min-h-[44px] min-w-[44px] flex items-center justify-center"
-        aria-label="Open menu"
-      >
-        <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
-        </svg>
-      </button>
-      <span className="text-[11px] text-text-muted">do-web-doc-resolver</span>
-    </div>
-    <Link href="/help" className="text-[11px] text-text-muted hover:text-accent min-h-[44px] flex items-center px-2">
-      Help
-    </Link>
-  </div>
-);
-
-interface SearchSectionProps {
-  query: string;
-  setQuery: (query: string) => void;
-  handleSubmit: (e?: React.FormEvent) => void;
-  loading: boolean;
-  inputRef: React.RefObject<HTMLInputElement | null>;
-  error: string;
-  hasResult: boolean;
-  onClear: () => void;
-  providerStatus: string | null;
-  isUrl: boolean;
-}
-
-const SearchSection = ({
-  query,
-  setQuery,
-  handleSubmit,
-  loading,
-  inputRef,
-  error,
-  hasResult,
-  onClear,
-  providerStatus,
-  isUrl,
-}: SearchSectionProps) => (
-  <div className="border-b-2 border-border-muted p-4">
-    <div className="flex items-center gap-4">
-      <label htmlFor="search-input" className="sr-only">
-        URL or search query
-      </label>
-      <div className="flex-1 relative">
-        <input
-          id="search-input"
-          ref={inputRef}
-          type="text"
-          value={query}
-          onChange={(e) => setQuery(e.target.value)}
-          onKeyDown={(e) => e.key === "Enter" && handleSubmit()}
-          placeholder="URL or search query..."
-          aria-invalid={Boolean(error)}
-          aria-errormessage={error ? "search-error" : undefined}
-          className="w-full bg-transparent text-[20px] sm:text-[24px] text-foreground placeholder:text-text-dim tracking-tight pr-10"
-        />
-        {query && (
-          <button
-            type="button"
-            onClick={() => {
-              setQuery("");
-              inputRef.current?.focus();
-            }}
-            className="absolute right-0 top-1/2 -translate-y-1/2 p-2 text-text-dim hover:text-accent transition-colors"
-            aria-label="Clear query"
-          >
-            ×
-          </button>
-        )}
-      </div>
-      {(query.trim() || hasResult) && (
-        <div className="flex items-center gap-2">
-          {query.trim() && (
-            <button
-              onClick={() => handleSubmit()}
-              disabled={loading}
-              aria-label={loading ? "Fetching results..." : "Fetch results"}
-              className="bg-accent text-background px-4 py-2 text-[13px] font-bold hover:bg-[#00cc33] disabled:opacity-50 min-w-[60px] min-h-[44px]"
-            >
-              {loading ? "..." : "Fetch"}
-            </button>
-          )}
-          <button
-            onClick={onClear}
-            aria-label="Clear input and results"
-            className="bg-transparent text-text-dim px-4 py-2 text-[13px] border-2 border-border-muted hover:border-accent hover:text-accent min-h-[44px]"
-          >
-            Clear
-          </button>
-        </div>
-      )}
-    </div>
-    {query.trim() && (
-      <div className="text-[11px] text-text-muted mt-2 uppercase tracking-wider">
-        {isUrl ? "Resolving as URL" : "Searching"}
-      </div>
-    )}
-    {providerStatus && (
-      <div role="status" aria-live="polite" className="text-[11px] text-accent mt-2 animate-pulse">
-        {providerStatus}
-      </div>
-    )}
-  </div>
-);
-
-interface MetadataBarProps {
-  sourceProvider: string | null;
-  resolveTime: number | null;
-  charCount: number;
-  qualityScore: number | null;
-  viewRaw: boolean;
-  setViewRaw: (view: boolean) => void;
-  handleCopyResult: () => void;
-  copied: boolean;
-}
-
-const MetadataBar = ({
-  sourceProvider,
-  resolveTime,
-  charCount,
-  qualityScore,
-  viewRaw,
-  setViewRaw,
-  handleCopyResult,
-  copied,
-}: MetadataBarProps) => (
-  <div className="flex items-center justify-between flex-wrap gap-3 px-4 py-2 border-b-2 border-border-muted text-[11px] text-text-muted">
-    <div className="flex items-center gap-4 flex-wrap">
-      <span>
-        Source: <span className="text-accent">{sourceProvider}</span>
-      </span>
-      {resolveTime && <span>{resolveTime}ms</span>}
-      <span>{charCount.toLocaleString()} chars</span>
-      {qualityScore !== null && (
-        <span title="Quality score (0-100)">
-          Quality:{" "}
-          <span
-            className={
-              qualityScore >= 70 ? "text-accent" : qualityScore >= 40 ? "text-[#ffaa00]" : "text-error"
-            }
-          >
-            {qualityScore}
-          </span>
-        </span>
-      )}
-    </div>
-    <div className="flex items-center gap-2">
-      <button
-        onClick={() => setViewRaw(false)}
-        className={`px-3 py-1 border border-border-muted ${!viewRaw ? "text-accent border-accent" : "text-text-muted"}`}
-        aria-pressed={!viewRaw}
-      >
-        Cards
-      </button>
-      <button
-        onClick={() => setViewRaw(true)}
-        className={`px-3 py-1 border border-border-muted ${viewRaw ? "text-accent border-accent" : "text-text-muted"}`}
-        aria-pressed={viewRaw}
-      >
-        Raw
-      </button>
-      <button
-        onClick={handleCopyResult}
-        aria-label={copied ? "Copied to clipboard" : "Copy to clipboard"}
-        aria-live="polite"
-        className="hover:text-foreground transition-colors min-h-[36px] px-2"
-      >
-        {copied ? "Copied" : "Copy"}
-      </button>
-    </div>
-  </div>
-);
 
 export const MainContent = (props: MainContentProps) => {
   const {
@@ -261,8 +83,6 @@ export const MainContent = (props: MainContentProps) => {
     inputRef.current?.focus();
   };
 
-  const charCount = result.length;
-
   return (
     <div id="main-content" className="flex-1 flex flex-col min-h-0" tabIndex={-1}>
       <MainHeader setMobileMenuOpen={setMobileMenuOpen} />
@@ -286,13 +106,13 @@ export const MainContent = (props: MainContentProps) => {
         </div>
       )}
 
-      <div className="flex-1 flex flex-col min-h-0">
+      <main className="flex-1 flex flex-col min-h-0">
         {result ? (
           <>
             <MetadataBar
               sourceProvider={sourceProvider}
               resolveTime={resolveTime}
-              charCount={charCount}
+              charCount={result.length}
               qualityScore={qualityScore}
               viewRaw={viewRaw}
               setViewRaw={setViewRaw}
@@ -324,7 +144,7 @@ export const MainContent = (props: MainContentProps) => {
             Paste a URL or enter a search query
           </div>
         )}
-      </div>
+      </main>
     </div>
   );
 };
