+++ web/app/components/MetadataBar.tsx
@@ -1,0 +1,69 @@
+"use client";
+
+interface MetadataBarProps {
+  sourceProvider: string | null;
+  resolveTime: number | null;
+  charCount: number;
+  qualityScore: number | null;
+  viewRaw: boolean;
+  setViewRaw: (view: boolean) => void;
+  handleCopyResult: () => void;
+  copied: boolean;
+}
+
+export const MetadataBar = ({
+  sourceProvider,
+  resolveTime,
+  charCount,
+  qualityScore,
+  viewRaw,
+  setViewRaw,
+  handleCopyResult,
+  copied,
+}: MetadataBarProps) => (
+  <div className="flex items-center justify-between flex-wrap gap-3 px-4 py-2 border-b-2 border-border-muted text-[11px] text-text-muted">
+    <div className="flex items-center gap-4 flex-wrap">
+      <span>
+        Source: <span className="text-accent">{sourceProvider}</span>
+      </span>
+      {resolveTime && <span>{resolveTime}ms</span>}
+      <span>{charCount.toLocaleString()} chars</span>
+      {qualityScore !== null && (
+        <span title="Quality score (0-100)">
+          Quality:{" "}
+          <span
+            className={
+              qualityScore >= 70 ? "text-accent" : qualityScore >= 40 ? "text-[#ffaa00]" : "text-error"
+            }
+          >
+            {qualityScore}
+          </span>
+        </span>
+      )}
+    </div>
+    <div className="flex items-center gap-2">
+      <button
+        onClick={() => setViewRaw(false)}
+        className={`px-3 py-1 border border-border-muted ${!viewRaw ? "text-accent border-accent" : "text-text-muted"}`}
+        aria-pressed={!viewRaw}
+      >
+        Cards
+      </button>
+      <button
+        onClick={() => setViewRaw(true)}
+        className={`px-3 py-1 border border-border-muted ${viewRaw ? "text-accent border-accent" : "text-text-muted"}`}
+        aria-pressed={viewRaw}
+      >
+        Raw
+      </button>
+      <button
+        onClick={handleCopyResult}
+        aria-label={copied ? "Copied to clipboard" : "Copy to clipboard"}
+        aria-live="polite"
+        className="hover:text-foreground transition-colors min-h-[36px] px-2"
+      >
+        {copied ? "Copied" : "Copy"}
+      </button>
+    </div>
+  </div>
+);
